+++ src/main/java/prgms/springbasic/voucher/VoucherService.java
@@ -1,34 +1,9 @@
 package prgms.springbasic.voucher;
 
-
-import prgms.springbasic.repository.VoucherRepository;
-
-import java.text.MessageFormat;
 import java.util.List;
 import java.util.UUID;
 
-public class VoucherService {
-    private final VoucherRepository voucherRepository;
-
-    public VoucherService(VoucherRepository voucherRepository) {
-        this.voucherRepository = voucherRepository;
-    }
-
-    public Voucher createFixedAmountVoucher(UUID voucherId, int amount) {
-        return new FixedAmountVoucher(voucherId, amount);
-    }
-
-    public Voucher createPercentDiscountVoucher(UUID voucherId, long percent) {
-        return new PercentDiscountVoucher(voucherId, percent);
-    }
-
-    public List<Voucher> getVoucherList() {
-        return voucherRepository.getVoucherList();
-    }
-
-    public Voucher getVoucher(UUID voucherId) {
-        return voucherRepository.findById(voucherId)
-                .orElseThrow(() -> new RuntimeException(MessageFormat.format("해당 바우처를 찾을 수 없습니다. 바우처 ID = {0}", voucherId))
-                );
-    }
+public interface VoucherService {
+    Voucher createVoucher(VoucherType voucherType, UUID voucherId, String value);
+    List<Voucher> getVoucherList();
 }
