+++ src/main/java/prgms/springbasic/voucher/VoucherType.java
@@ -1,0 +1,6 @@
+package prgms.springbasic.voucher;
+
+public enum VoucherType {
+    FIXEDAMOUNTVOUCHER,
+    PERCENTDISCOUNTVOUCHER;
+}
